==> 02_CSV.ipynb <==
import csv
import os
import pandas as pd

### Create CSV and fill in with images info
Pandas (but very, very long)

# create df
data_df = pd.DataFrame(columns=['name', 'path', 'label'])

# add new df row for each image
for image in os.listdir('/Users/hela/Code/pata/spectrograms/'):
    new_row = pd.DataFrame(columns=[image, os.path.join('/Users/hela/Code/pata/data_test.csv', image), ''])
    data_df = pd.concat([data_df, new_row], ignore_index=True)
data_df.to_csv('/Users/hela/Code/pata/data_pd.csv', index=False)

CSV

# create CSV with three columns
with open('/Users/hela/Code/pata/data_test.csv', 'w', newline='') as file:  # write
    csv.writer(file).writerow(['name', 'path', 'label'])

# add new row for each image
with open('/Users/hela/Code/pata/data_test.csv', 'a', newline='') as file:  # append
    for image in os.listdir('/Users/hela/Code/pata/spectrograms/'):
        csv.writer(file).writerow([image, os.path.join('/Users/hela/Code/pata/spectrograms/', image), ''])

# open CSV with pd
data_df = pd.read_csv('/Users/hela/Code/pata/data_test.csv')
data_df

### Image + audio info

with open('/Users/hela/Code/pata/data_test.csv', 'a', newline='') as file:
    csv.writer(file).writerow(['name', 'audio_name', 'audio_path', 'image_name', 'image_path', 'label'])
    major_dir = '/Users/hela/Code/pata/Audio_data/'
    for folder in os.listdir(major_dir):
        dir = major_dir + folder + '/'
        for audio in os.listdir(dir):
            image_name = audio.split('.')[0] + '.png'
            csv.writer(file).writerow([audio.split('.')[0],
                                       audio,
                                       os.path.join(dir, audio),
                                       image_name,
                                       os.path.join('/Users/hela/Code/pata/spectrograms/', image_name),
                                       ''])

data_df = pd.read_csv('/Users/hela/Code/pata/data_test.csv')
data_df
