==> 04_Labeling.ipynb <==
from ipywidgets import widgets
from IPython.display import Audio, display, Image, clear_output
import pandas as pd

data_df = pd.read_csv('/Users/hela/Code/pata/data_labeling.csv')

row_index = None

def set_new_trial():
    global row_index, unlabeled_trials
    unlabeled_trials = data_df[data_df['label'].isna()]
    rand_row = unlabeled_trials.sample(n=1)
    row_index = rand_row.index[0]
    path_audio = rand_row['audio_path'].iloc[0]
    path_img = rand_row['image_path'].iloc[0]
    print(rand_row['name'].iloc[0])
    display(Image(path_img))
    display(Audio(path_audio, autoplay=True))

# Widgets
output = widgets.Output()
button_pa = widgets.Button(description = 'PA')
button_ta = widgets.Button(description = 'TA')
button_err = widgets.Button(description = 'Broken')
button_next = widgets.Button(description = '->')
button_stop = widgets.Button(description = 'Save progress')
display(widgets.HBox([button_pa, button_ta, button_err, button_next]))
display(widgets.HBox([output]))
display(widgets.HBox([button_stop]))

def update_display():
    with output: 
        clear_output()
        set_new_trial()

def click_pa(_):
    with output:
        data_df.at[row_index, 'label'] = 'pa'
        update_display()
button_pa.on_click(click_pa)

def click_ta(_):
    with output:
        data_df.at[row_index, 'label'] = 'ta'
        update_display()
button_ta.on_click(click_ta)

def click_err(_):
    with output:
        data_df.at[row_index, 'label'] = 'err'
        update_display()
button_err.on_click(click_err)

def click_next(_):
    with output:
        update_display()
button_next.on_click(click_next)

def click_stop(_):
    with output:
        data_df.to_csv('/Users/hela/Code/pata/data_labeling.csv', index=False)
        print('Updates are saved.')
        print(data_df['label'].notna().sum(), '/', 8640, 'labeled trials.')
        print(data_df['label'].value_counts())
button_stop.on_click(click_stop)
    
# Start
update_display()

# check that there are new values in the 'label' column
df = pd.read_csv('/Users/hela/Code/pata/data_labeling.csv')
df['label'].value_counts()  # counts of unique values

# check that there are new values in the 'label' column
df = pd.read_csv('/Users/hela/Code/pata/data_labeling.csv')
df[df['label'].notna()]  # examine rows with selected values
