==> 05_Labeled_and_unlabeled_sets.ipynb <==
import pandas as pd

# load the labeling csv
df = pd.read_csv('/Users/hela/Code/pata/data_labeling.csv')
df['label'].value_counts()

# Extract labeles from labeling csv

# select only labeled rows:
df_labels = df[df['label'].notna()]
# select only rows with pa and ta:
df_without_err = df_labels[df_labels['label']!='err']
# # rename categories into numbers:
#df_without_err['label'] = df_without_err['label'].map({'pa':0, 'ta':1})
# check:
print(df_without_err['label'].value_counts())
df_without_err

# save the file:
df_without_err.to_csv('/Users/hela/Code/pata/data_labeled.csv', index=False)

# Add labeles from predicted & checked table

# load already labeled csv
df_labeled = pd.read_csv('/Users/hela/Code/pata/data_labeled.csv')
print(df_labeled['label'].value_counts())
df_labeled

# load predicted & checked csv
df_predicted = pd.read_csv('/Users/hela/Code/pata/data_predicted.csv')
# modify by removing errs, predict columns, na
df_predicted = df_predicted[df_predicted['label']!='err']
df_predicted = df_predicted.drop(['predict', 'confidence'], axis=1)
df_predicted = df_predicted[df_predicted['label'].notna()]
print(df_predicted['label'].value_counts())
df_predicted

# merged csv (vertical stack)
df_labeled_new = df_labeled._append(df_predicted)
print(df_labeled_new['label'].value_counts())
df_labeled_new

# save extended label csv:
df_labeled_new.to_csv('/Users/hela/Code/pata/data_labeled.csv', index=False)
